Augimas v0.0252 - Bug Fixes. More Bugs to Solve.
Fixes:
- New User Tapping Back Button while in the TeamManagement Screen
(Before they made an account, causes the TeamManagement Screen to
disappear) but no Settings Screen to appear.
- User entering their chat section cannot input text in the Chat, but
they can in the Host's section.
- User entering their chat section can't see start the conversation on
either channels.
- User Creating new team and joining it doesn’t release proper
notifications.
- Join & Left Team Notifications were being released 2x as many as it
was supposed to.
- Users permission gets set to what it was previously upon tapping the
update team name/username fields in the Team Management Screen (sends
notification).
————————————————
New Bugs:
- The Client/Host Chat Input doesn't rise above the keyboard
- Now when the keyboard is focused on the Team Name/Username Input
Fields in the Team Management Screen, it takes two modifications of a
User’s Role for it to send a notification.
————————————————
Old Bugs:
- Smooth Transitions from the Settings Fragment to the Team Management
Fragment (it shows the previous Fragment that led to the Settings
Fragment for a brief second)