Augimas v0.0252 -> v0.0253: Bug Fixes & New Bugs.
----------------------- New Bugs ----------------------
=== Notifications Fragment
- A Client team is able to see the Notification that goes to them as
well as the Notification that goes to the Host team.
- Adding a Branding Item sends more than 2x the Number of Notifications
it is supposed to
===  Branding Name Fragment
- Tapping the "x" button on a new empty item then directs attention to
the first item's input field.
- Tapping the "x" button on a filled item occasionally causes an item
to switch the text back and forth between its name and the previous
item's name over and over, causing a pileup of notifications.
=== Branding Name Fragment
- Tapping the "+" Button on a new Branding Name Item causes the rest of
the unaffected Items to repopulate which causes a lot more visual
action than is necessary.
- Updating a Branding Name Item shouldn't say that the User "added" it.
It should say what it was changed from to what it was changed to.
=== Notifications
- New Message Notifications should say (when coming from another Team)
"[User Name] from [their Team Name] said, '[Quote]'"
=== Team Management Fragment
- Inviting a User that is already in your Team lets you know that they
are in your team, not just a general team.
--------------------------------------------------------

----------------------- Old Bugs -----------------------
=== Chat Fragment
- The Client/Host Chat Input doesn't rise above the keyboard
- Now when the keyboard is focused on the Team Name/Username Input
Fields in the Team Management Screen, it takes two modifications of a
User’s Role for it to send a notification.
=== Team Management Fragment
- Smooth Transitions from the Settings Fragment to the Team Management
Fragment (it shows the previous Fragment that led to the Settings
Fragment for a brief second)
--------------------------------------------------------

---------------------- Fixed Bugs ----------------------
=== Recent Activities Fragment
- Entity Role text added to the end of a Notification doesn't include a
"."
- Action Bar Title for the Recent Activities Screen should say "[Team
Name]"
- Renamed Notification Navigation Item to "Recent Activities"
=== Branding Name Fragment
- Tapping the "x" Button on a Branding Name Item that is filled causes
it to go through lots of visual changes that are unnecessary, just get
rid of the view.
- Action Bar Title for the Branding Element Screens should say, "[Team
Name]'s Domain Names" or "[Team Name]'s Social Media Names"
appropriately.
- Tapping the "x" Button on a Branding Name Item causes the button to
rotate outside of it's bounds. Should fix so that it rotates first,
then expands in size.
=== Branding Elements Fragment
- Action Bar Title for the Branding Elements Screen should say "[Team
Name]'s Branding Elements"
=== Team Management Fragment
- Action Bar Title for the Team Management Fragment should say "Team
[Team Name]"
--------------------------------------------------------