Augimas v0.02535 -> v0.0254: Bug Fixes
--------------------- Fixed Bugs ----------------------
=== Team Management Fragment
- The User Role Spinner Selection is off by 1 item, displays Default->
Admin instead of None-> Admin
- Now when the keyboard is focused on the Team Name/Username Input
Fields in the Team Management Screen, it takes two modifications of a
User’s Role for it to send a notification (though it updates the data).
- You are able to select a user with the same role as yourself
(shouldn't be able to), but you cannot modify it.
=== Recent Activities Fragment
- Changed Notification Class to-> Recent Activity Class for clarity
- Refactoring Notification Class to Recent Activity Class caused all
instances of Notification to change to Recent Activity including the
firebase reference which caused it to break off from all other
notifications. Had to export JSON and rename notifications to
recentActivities.
- Doesn't direct users fully to the location in the application.
=== Branding Elements & Branding Element Fragments
- Users with lower than a Editor Role should not be able to add/update
new Branding Items or update the Status of a Branding Element
=== Chat Fragment
- The Client/Host Chat Input doesn't rise above the keyboard
- RecentActivity won't direct you to the specific channel
- Chat Activities don't correctly state who the message came from when
sent from median channel.
- Tapping the Chat Activity from your Team doesn't properly display
either channel's messages.
- Tapping the Chat Activity from "the other" Team properly displays the
shared channel, but not your teams messages.
- Excessive space at the end of a message doesn't get clipped.
- Chat Message now effectively states to what team it was going to.
- Tapping any Client Chat Activity doesn't properly name the Tabs. They
are both named to the Client's Team
- The Client-side Channel Doesn't Display. It is throws you in a
foreign chat.
--------------------------------------------------------