Augimas v0.02552 -> v0.02553
-------------------- Fixed Bugs -----------------------
- Sign Up: See the bottom right, the "Or" is crunched up against the
"Sign In" so it says "OrSign In"  I see that this looked fine in your
video.
- You will want to allow users to change their password since many will
change their email address at times, for instance if changing from a
gmail.com to a corporate domain email address.  The way the app is
setup, this will not be possible. 
- You will want to allow a user to change full name. If they have a
typo during sign up, there will be no other way for them to change.
- The tab menu is not changing when clicking on an items in the "Recent
Activities".  I clicked on "Augimas updated Drew Brighton's role to
Owner" and goes to the setting area but the tab menu stays on "Recent
Activities." The menu is not really showing the true location of the
app.  The video walkthrough shows this as well.
--------------------------------------------------------