Augimas v0.0255545
Changes:
-The "x" button immediately turns into a "+" when your input is valid
and is able to be stored (previously it would turn the moment you
inputted any data whether it was valid or not).
- The logo section still had a post method or two for displaying some
UI stuff so I had those removed to ensure that it wouldn't prevent UI
from showing on time. - Changed the Social Media Section input to force
place an "@" at the beginning of the input that cannot be removed.
- If you try to access a feature when you don't have the data it will
let you go to that screen as it should, it just doesn't display
anything until you get an internet connection and refresh the screen.